move Measurement and Series to inmem package
